GraphLaplacians

GraphLaplacians.degree_matrix

GraphLaplacians.degrees

GraphLaplacians.inv_sqrt_degree_matrix

GraphLaplacians.laplacian_matrix

GraphLaplacians.normalized_laplacian

GraphLaplacians.random_walk_laplacian

GraphLaplacians.scaled_laplacian

GraphLaplacians.signless_laplacian

GraphLaplacians.degree_matrix — Function

degree_matrix(g[, T]; dir=:out)

Degree matrix of graph g. Return a matrix which contains degrees of each vertex in its diagonal. The values other than diagonal are zeros.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

Examples

julia> using GraphLaplacians

julia> m = [0 1 1; 1 0 0; 1 0 0];

julia> GraphLaplacians.degree_matrix(m)
3×3 LinearAlgebra.Diagonal{Int64, Vector{Int64}}:
 2  ⋅  ⋅
 ⋅  1  ⋅
 ⋅  ⋅  1

GraphLaplacians.degrees — Method

degrees(g[, T]; dir=:out)

Degree of each vertex. Return a vector which contains the degree of each vertex in graph g.

Arguments

g: should be a adjacency matrix, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

Examples

julia> using GraphLaplacians

julia> m = [0 1 1; 1 0 0; 1 0 0];

julia> GraphLaplacians.degrees(m)
3-element Vector{Int64}:
 2
 1
 1

GraphLaplacians.inv_sqrt_degree_matrix — Function

inv_sqrt_degree_matrix(g[, T]; dir=:out)

Inverse squared degree matrix of graph g. Return a matrix which contains inverse squared degrees of each vertex in its diagonal. The values other than diagonal are zeros.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

GraphLaplacians.laplacian_matrix — Function

laplacian_matrix(g[, T]; dir=:out)

Laplacian matrix of graph g.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

GraphLaplacians.normalized_laplacian — Function

normalized_laplacian(g[, T]; selfloop=false)

Normalized Laplacian matrix of graph g.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

selfloop: adding self loop while calculating the matrix (optional).

GraphLaplacians.random_walk_laplacian — Function

random_walk_laplacian(g[, T]; dir=:out)

Random walk normalized Laplacian matrix of graph g.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).

GraphLaplacians.scaled_laplacian — Function

scaled_laplacian(g[, T])

Scaled Laplacien matrix of graph g, defined as $\hat{L} = \frac{2}{\lambda_{max}} L - I$ where $L$ is the normalized Laplacian matrix.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

GraphLaplacians.signless_laplacian — Function

signless_laplacian(g[, T]; dir=:out)

Signless Laplacian matrix of graph g.

Arguments

g: should be a adjacency matrix, FeaturedGraph, SimpleGraph, SimpleDiGraph (from LightGraphs) or SimpleWeightedGraph, SimpleWeightedDiGraph (from SimpleWeightedGraphs).

T: result element type of degree vector; default is the element type of g (optional).

dir: direction of degree; should be :in, :out, or :both (optional).